+++ PayPal/src/main/java/com/braintreepayments/api/paypal/PayPalVaultRequest.kt
@@ -25,16 +25,14 @@ import org.json.JSONObject
  * @see [Examples of prominent in-app disclosures](https://support.google.com/googleplay/android-developer/answer/9799150?hl=en.Prominent%20in-app%20disclosure)
  *
  * @property shouldOfferCredit Offers PayPal Credit if the customer qualifies. Defaults to false.
- * @property recurringBillingDetails Optional: Recurring billing product details.
- * @property recurringBillingPlanType Optional: Recurring billing plan type, or charge pattern.
  */
 @Parcelize
 class PayPalVaultRequest
 @JvmOverloads constructor(
     override val hasUserLocationConsent: Boolean,
     var shouldOfferCredit: Boolean = false,
-    var recurringBillingDetails: PayPalRecurringBillingDetails? = null,
-    var recurringBillingPlanType: PayPalRecurringBillingPlanType? = null,
+    override var recurringBillingDetails: PayPalRecurringBillingDetails? = null,
+    override var recurringBillingPlanType: PayPalRecurringBillingPlanType? = null,
     override var enablePayPalAppSwitch: Boolean = false,
     override var localeCode: String? = null,
     override var billingAgreementDescription: String? = null,
@@ -47,7 +45,8 @@ class PayPalVaultRequest
     override var riskCorrelationId: String? = null,
     override var userAuthenticationEmail: String? = null,
     override var userPhoneNumber: PayPalPhoneNumber? = null,
-    override var lineItems: List<PayPalLineItem> = emptyList()
+    override var lineItems: List<PayPalLineItem> = emptyList(),
+    override var userAction: PayPalPaymentUserAction = PayPalPaymentUserAction.USER_ACTION_DEFAULT
 ) : PayPalRequest(
     hasUserLocationConsent = hasUserLocationConsent,
     localeCode = localeCode,
@@ -62,6 +61,7 @@ class PayPalVaultRequest
     enablePayPalAppSwitch = enablePayPalAppSwitch,
     userAuthenticationEmail = userAuthenticationEmail,
     lineItems = lineItems,
+    userAction = userAction,
 ) {
 
     @OptIn(ExperimentalBetaApi::class)
@@ -116,6 +116,10 @@ class PayPalVaultRequest
 
         if (localeCode != null) {
             experienceProfile.put(LOCALE_CODE_KEY, localeCode)
+        }
+
+        if (userAction != PayPalPaymentUserAction.USER_ACTION_DEFAULT) {
+            experienceProfile.put(USER_ACTION_KEY, userAction.stringValue)
         }
 
         if (shippingAddressOverride != null) {