+++ CHANGELOG.md
@@ -1,4 +1,15 @@
 # Braintree Android SDK Release Notes
+  
+## unreleased
+
+* PayPal
+  * Add `recurringBillingDetails`, `recurringBillingPlanType`, and `amountBreakdown` properties to `PayPalCheckoutRequest`. Enables RBA metadata to be passed for the PayPal Checkout Vault with Purchase flow
+  * Add `userAction` property to `PayPalVaultRequest`
+
+## unreleased
+
+* BraintreeCore
+    * Add OkHttp version 4.12.0
 
 ## 5.17.0 (2025-09-09)
 
@@ -14,6 +25,7 @@
   * Fix an issue where `PayPalRequest` was sending `phone_number` instead of `payer_phone`
   * Add `merchant` and `flow_type` as query parameters to the app switch URL.
   * Add `paymentId` to `PayPalAccountNonce`
+
 * GooglePay
     * Add `softwareInfo` details to the `merchantInfo` field in the `loadPaymentData` request.